+++ frontend/main/src/RarlaTodo/components/TodoBox/index.scss
@@ -1,0 +1,8 @@
+.todo-box {
+  box-sizing: border-box;
+  width: 600px;
+  height: 600px;
+  padding: 10px;
+  border: 4px solid black;
+  font-size: 20px;
+}
