+++ frontend/main/src/RarlaTodo/index.scss
@@ -4,17 +4,6 @@
   align-items: center;
   flex-direction: column;
   
-  
-  
-  .todo-box {
-    box-sizing: border-box;
-    width: 600px;
-    height: 600px;
-    padding: 10px;
-    border: 4px solid black;
-    font-size: 20px;
-  }
-  
   ul, li {
     margin: 0;
     padding: 0;
